Fixed size of Atom type on 64 Bit systems, closes #4
Atom is always a unsigned long, and not just 32 bit long

+++ synaptiks/_bindings/xlib.py
@@ -48,7 +48,7 @@ from synaptiks._bindings.util import load_library, scoped_pointer
 c_ubyte_p = POINTER(c_ubyte)
 
 # X11 types
-Atom = c_uint32
+Atom = c_ulong
 Atom_p = POINTER(Atom)
 Bool = c_int
 Status = c_int
